Clean build time on my slow laptop (this will not be comparable to yours, but for me it's reproducible across builds)

|Before|With ccache|With ccache and linker selection|
|--|--|--|
|48.3s|8.5s|5.6s (ld.gold)<br>5.5s (ld.lld)|


The goal here is not only to improve ccache compile times, but also to provide a template to others how to improve compilation times.