@jrosdahl 2021-02-01
I noticed that the overhead of `cmake -E env` is ≈9 ms on my system. This means that ccache direct mode hits on average have become about twice as slow when building ccache itself on my system.

I improved this in 68d8481c173eb711624232c8f9c8e4507d747dc9 by using the standard `env` program if available. Its overhead is ≈1 ms.